Moved concern in separate file
