+++ lib/pluck_to_hash/railtie.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
-require 'pluck_to_hash'
-
 module PluckToHash
   class Railtie < Rails::Railtie
     initializer 'pluck_to_hash.initialization' do
-      ActiveSupport.on_load(:active_record) { include PluckToHash }
+      ActiveSupport.on_load(:active_record) {
+        require 'pluck_to_hash/pluck_to_concern'
+        include PluckToConcern
+      }
     end
   end
 end
